Update - Prevent url edit video, delete video

+++ controllers/videoController.js
@@ -79,7 +79,6 @@ export const videoDetail = async (req, res) => {
 	} = req;
 	try {
 		const video = await Video.findById(id).populate("creator");
-		console.log(video);
 		res.render("videoDetail", {
 			pageTitle: video.title,
 			video
@@ -97,10 +96,15 @@ export const getEditVideo = async (req, res) => {
 	} = req;
 	try {
 		const video = await Video.findById(id);
-		res.render('editVideo', {
-			pageTitle: 'editVideo',
-			video,
-		});
+		// console.log(req.user.id)
+		if (video.creator !== req.user.id) {
+			throw Error();
+		} else {
+			res.render('editVideo', {
+				pageTitle: 'editVideo',
+				video,
+			});
+		}
 	} catch (error) {
 		res.redirect(routes.home);
 	}
@@ -137,9 +141,14 @@ export const deleteVideo = async (req, res) => {
 		},
 	} = req;
 	try {
-		await Video.findByIdAndRemove({
-			_id: id,
-		});
+		const video = await Video.findById(id);
+		if (video.creator !== req.user.id) {
+			throw Error();
+		} else {
+			await Video.findByIdAndRemove({
+				_id: id,
+			});
+		}
 	} catch (error) {
 		console.log(error);
 	}
